6 add test code from xml generator (#8)
* Add generated xml tests and fix integer types

* Fix use statements

* Make ChanIndex vec

* Represent 130 polarization regex types with String

* Add tests, make some things optinal

* Fix use statment

* Format

* Add gen xml 040 test

* Add more gen xml tests

* Fix integer type

* Fix enums in test data

* Make RefTime optional

* Create 040 RMAlgo

* Silence unused import false alarm

* Remove extra conversion

* Remove dup code from 050 image_formation.rs

* Move v130 gen xml tests

* Move v050 gen xml test, re-export top lev

* Move gen xml v040 tests

* Move gen xml grid 040 tests

* Revert some types to usize

* Add schemas

* Add schema copies to modify

* Modify schema exp/order to unsigned

* Update 130 gen xml test

* Add smarter gen 130 tests

* Improve hybrid test

* Fix use statement

* Fix schema enums

* Rename ScpCoa to match tag SCPCOA

* Rename Pfa to PFA to match tag

* Add generated 050 xml data

* Complete component test, add 050 test

* Rename test

* Remove old tests

* Add 050 component test

* Make Corner optional

* Add match info test

* Handle attributes in gen xml

* Handle attributes

* Rename pre 1.0 RgAzComp to RGAZCOMP to match key

* Add generated 040 tests

* Run cargo fmt

* Rename Rma to RMA to match key

* Add another 130 xml dataset

* Make dep ChanParameters optional

* Add another gen 050 test

* Fix RGAZCOMP name in dep

* Add more tests, missing RgAzComp

* Add rg_az_comp parsing tests

* Make size u64 everywhere

* Change int index attr to usize

* Change Num/First,Row/Col to unsigned

* Make num_chan_proc unsigned

* Make ChanIndex unsigned

* Make rcv_apc_index unsigned

* Make num lines/samples unsigned

* Make num_match_types unsigned

* Make num match collections unsigned

* Revert "Represent 130 polarization regex types with String"

This reverts commit f39c25463f48f6704972f2211ab6d11502485ca7.

* Add missing serde other proc macros

* Update tests/parse_gen_xml.rs


---------

+++ src/dep/v0_4_0/image_data.rs
@@ -12,9 +12,9 @@ pub struct ImageData {
     #[serde(rename = "NumCols")]
     pub num_cols: u64,
     #[serde(rename = "FirstRow")]
-    pub first_row: u64,
+    pub first_row: usize,
     #[serde(rename = "FirstCol")]
-    pub first_col: u64,
+    pub first_col: usize,
     #[serde(rename = "FullImage")]
     pub full_image: FullImage,
     #[serde(rename = "SCPPixel")]
@@ -40,14 +40,14 @@ pub enum PixelTypeEnum {
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct AmpTable {
     #[serde(rename = "@size")]
-    pub size: u16, // 256
+    pub size: u64, // 256
     #[serde(rename = "Amplitude")]
     pub amplitude: Vec<Amplitude>,
 }
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct Amplitude {
     #[serde(rename = "@index")]
-    pub index: u8, // [0, 255]
+    pub index: usize, // [0, 255]
     #[serde(rename = "$value")]
     pub value: f64,
 }
